исправляет замечания по 7 заданию
