+++ js/main.js
@@ -1,10 +1,12 @@
 import { getOffers } from './data.js';
 import { createPopup } from './popup.js';
 import { addPopup } from './map.js';
-import { turnActive } from './form.js';
+import { turnFormOn } from './form.js';
+import { turnFiltersOn } from './map-filters.js';
 
 const popups = Array.from(getOffers(), (item) => createPopup(item));
 
 addPopup(popups[0]);
 
-turnActive();
+turnFormOn();
+turnFiltersOn();
